feat(mobile-navbar): abstract scroll effect
