+++ package.json
@@ -15,13 +15,14 @@
     "@types/react": "^17.0.38",
     "@types/react-dom": "^17.0.11",
     "framer-motion": "^5.6.0",
+    "hamburger-react": "^2.5.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-icons": "^4.3.1",
     "react-pdf": "^5.6.0",
     "react-reveal": "^1.2.2",
     "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
+    "react-scripts": "^2.1.3",
     "styled-components": "^5.3.3",
     "typescript": "^4.5.4",
     "web-vitals": "^2.1.3"
