+++ src/hooks/useScroll.tsx
@@ -1,0 +1,22 @@
+import { useState, useEffect } from "react";
+
+const useScroll = () => {
+    const [scrolled, setScrolled] = useState<boolean>(false);
+
+    useEffect(() => {
+        const handleScroll = () => {
+            const isScrolled = window.scrollY > 10;
+            if (isScrolled !== scrolled) {
+                setScrolled(!scrolled);
+            }
+        };
+        document.addEventListener("scroll", handleScroll, { passive: true });
+        return () => {
+            document.removeEventListener("scroll", handleScroll);
+        };
+    }, [scrolled]);
+
+    return scrolled;
+};
+
+export { useScroll };
